PALLADIO-530 move StoEx dialog to editor commons
