+++ bundles/org.palladiosimulator.pcm.stoex.ui/src/org/palladiosimulator/pcm/stoex/ui/services/StoexEditedResourceProvider.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.pcm.stoex.ui.dialog.impl;
+package org.palladiosimulator.pcm.stoex.ui.services;
 
 import java.util.UUID;
